batch_analyser/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
flashtext
